requirements.txt
beautifulsoup4==4.10.0
google-cloud-storage==1.42.3
icalendar==4.0.8
pytz==2021.3
requests==2.26.0
urllib3==1.26.7
